# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/owners/{ownerId}": {
      "get": {
        "description": "Retrieve owner information and associated cafes",
        "operationId": "OwnerController_getOwnerById",
        "parameters": [
          {
            "name": "ownerId",
            "required": true,
            "in": "path",
            "description": "Owner UUID",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOwnerByIdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized - invalid or missing JWT token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "code": {
                      "type": "string",
                      "example": "NOT_FOUND"
                    },
                    "message": {
                      "type": "string",
                      "example": "Resource not found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get owner by ID",
        "tags": [
          "owners"
        ]
      },
      "put": {
        "description": "Update owner information and associated cafes",
        "operationId": "OwnerController_updateOwner",
        "parameters": [
          {
            "name": "ownerId",
            "required": true,
            "in": "path",
            "description": "Owner UUID",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOwnerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerResponseWithTokens"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "code": {
                      "type": "string",
                      "example": "BAD_REQUEST"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid request data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized - invalid or missing JWT token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "code": {
                      "type": "string",
                      "example": "NOT_FOUND"
                    },
                    "message": {
                      "type": "string",
                      "example": "Resource not found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update owner",
        "tags": [
          "owners"
        ]
      },
      "delete": {
        "description": "Delete owner account and associated data",
        "operationId": "OwnerController_deleteOwner",
        "parameters": [
          {
            "name": "ownerId",
            "required": true,
            "in": "path",
            "description": "Owner UUID",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOwnerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized - invalid or missing JWT token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "code": {
                      "type": "string",
                      "example": "NOT_FOUND"
                    },
                    "message": {
                      "type": "string",
                      "example": "Resource not found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete owner",
        "tags": [
          "owners"
        ]
      }
    },
    "/api/owners": {
      "post": {
        "description": "Register a new cafe owner account",
        "operationId": "OwnerController_createOwner",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOwnerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Owner created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOwnerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "code": {
                      "type": "string",
                      "example": "BAD_REQUEST"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid request data"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 409
                    },
                    "code": {
                      "type": "string",
                      "example": "CONFLICT"
                    },
                    "message": {
                      "type": "string",
                      "example": "Owner with email already exists"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create new owner",
        "tags": [
          "owners"
        ]
      }
    },
    "/api/owners/login": {
      "post": {
        "description": "Authenticate owner and receive access tokens",
        "operationId": "OwnerController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginOwnerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerResponseWithTokens"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized - invalid or missing JWT token"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Login owner",
        "tags": [
          "owners"
        ]
      }
    }
  },
  "info": {
    "title": "Cafe IT API",
    "description": "API for managing cafes and seat availability",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "GetOwnerByIdResponse": {
        "type": "object",
        "properties": {
          "ownerId": {
            "type": "string",
            "description": "Unique identifier of the cafe owner",
            "example": "550e8400-e29b-41d4-a716-446655440001",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Owner email address",
            "example": "owner@example.com"
          },
          "cafeIds": {
            "description": "Array of cafe UUIDs owned by this owner",
            "example": [
              "550e8400-e29b-41d4-a716-446655440000",
              "550e8400-e29b-41d4-a716-446655440001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ownerId",
          "email",
          "cafeIds"
        ]
      },
      "CreateOwnerRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Owner email address",
            "example": "owner@example.com"
          },
          "password": {
            "type": "string",
            "description": "Owner password",
            "example": "securePassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateOwnerResponse": {
        "type": "object",
        "properties": {
          "ownerId": {
            "type": "string",
            "description": "Unique identifier of the cafe owner",
            "example": "550e8400-e29b-41d4-a716-446655440001",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Owner email address",
            "example": "owner@example.com"
          },
          "cafeIds": {
            "description": "Array of cafe UUIDs owned by this owner",
            "example": [
              "550e8400-e29b-41d4-a716-446655440000",
              "550e8400-e29b-41d4-a716-446655440001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessToken": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expiresIn": {
            "type": "number",
            "description": "Token expiration time in seconds",
            "example": 900
          }
        },
        "required": [
          "ownerId",
          "email",
          "cafeIds",
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "LoginOwnerRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Owner email address",
            "example": "owner@example.com"
          },
          "password": {
            "type": "string",
            "description": "Owner password",
            "example": "securePassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "OwnerResponseWithTokens": {
        "type": "object",
        "properties": {
          "ownerId": {
            "type": "string",
            "description": "Unique identifier of the cafe owner",
            "example": "550e8400-e29b-41d4-a716-446655440001",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Owner email address",
            "example": "owner@example.com"
          },
          "cafeIds": {
            "description": "Array of cafe UUIDs owned by this owner",
            "example": [
              "550e8400-e29b-41d4-a716-446655440000",
              "550e8400-e29b-41d4-a716-446655440001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessToken": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expiresIn": {
            "type": "number",
            "description": "Token expiration time in seconds",
            "example": 900
          }
        },
        "required": [
          "ownerId",
          "email",
          "cafeIds",
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "UpdateOwnerRequest": {
        "type": "object",
        "properties": {
          "cafeIds": {
            "description": "Array of cafe UUIDs owned by this owner",
            "example": [
              "550e8400-e29b-41d4-a716-446655440000",
              "550e8400-e29b-41d4-a716-446655440001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "cafeIds"
        ]
      },
      "DeleteOwnerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the deletion was successful",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      }
    }
  }
}